+++ madsim-tonic/src/client.rs
@@ -191,6 +191,18 @@ impl<F: Interceptor> Grpc<crate::transport::Channel, F> {
         }
         Ok(())
     }
+
+    /// Limits the maximum size of a decoded message.
+    pub fn max_decoding_message_size(mut self, _limit: usize) -> Self {
+        // self.config.max_decoding_message_size = Some(limit);
+        self
+    }
+
+    /// Limits the maximum size of an encoded message.
+    pub fn max_encoding_message_size(mut self, _limit: usize) -> Self {
+        // self.config.max_encoding_message_size = Some(limit);
+        self
+    }
 }
 
 async fn with_timeout<T>(