+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## tonic [0.3.1] - 2023-07-24
+
+### Added
+
+- tonic: Add dummy API `max_encoding_message_size` and `max_decoding_message_size`.
+
 ## tonic, etcd [0.3.0] - 2023-07-20
 
 ### Changed